+++ src/utils/cellSetOperations.js
@@ -1,3 +1,5 @@
+import * as setOperations from 'utils/setOperations';
+
 /**
  *
  * @param {*} cellClassKey The key of the cell class we want to unite by, e.g.: 'louvain'
@@ -84,9 +86,21 @@ const complement = (listOfSets, properties) => {
   return complementSet;
 };
 
+const withoutFilteredOutCells = (cellSets, originalCellIds) => {
+  const louvainClusters = cellSets.find(({ key }) => key === 'louvain').children;
+
+  const filteredInCellIds = louvainClusters.reduce(
+    (filteredInCellIdsAcum, { cellIds }) => setOperations.union(filteredInCellIdsAcum, cellIds),
+    new Set(),
+  );
+
+  return setOperations.intersection(filteredInCellIds, originalCellIds);
+};
+
 export {
   union,
   intersection,
   complement,
   unionByCellClass,
+  withoutFilteredOutCells,
 };
